examples/knowledge-graph/pyproject.toml
[project]
name = "knowledge-graph"
version = "0.1.0"
description = "This example covers ingestion of PDF files into a knowledge graph, and a chat UI to ask questions to an AI agent."
readme = "README.md"
authors = [
    { name = "Martin Schaer", email = "martin.schaer@surrealdb.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.55.1",
    "fastapi[standard]>=0.119.1",
    "kaig",
    "kreuzberg[keywords-yake]>=4.2.0",
    "logfire",
    "pydantic-ai-slim[openai]>=1.47.0",
    "python-multipart>=0.0.20",
    "tiktoken>=0.12.0",
    "surrealfs-ai"
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
kaig = { workspace = true }
surrealfs-ai = { git = "https://github.com/surrealdb/surrealfs", branch = "main", subdirectory = "python/surrealfs_ai" }
# surrealfs-ai = { path = "../../../surrealfs/python/surrealfs_ai", editable = true }

[tool.basedpyright]
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportUnknownMemberType = false
